fix: stop postgres service
Add missing `stop` task. 
When running `--tags=setup-all` or something equivalent this would cause the `postgres` service to stay up and running. I'm not able to test all the 'import' cases in the other util files so there might be other tree-paths out there that leave a dangling process. Maybe it's even better to add this task to the `main.yml`?

Anyway, need to ensure all process are killed when the user didn't ask for them to stay up.